Add extraction, protection and purification scrolls; allow decimal book chances
Scrolls are dragged onto an item like books. Extraction takes a random
removable enchant off and hands it back as a book; purification takes a curse
off; both roll their own success chance. Protection keeps the item on death
and saves it from a book that would destroy it, and is used up once it does.
All three are sold in /enchanter and given with /arkenchants scroll.

Book success and destroy chances are now decimals, so a fabled book can roll
0.1%, and group ranges in config.yml accept values like 0.1-100. Books made by
earlier versions still read their whole-number chances.

+++ arkenchants/src/main/java/com/arkcronist/enchants/model/Group.java
@@ -2,7 +2,7 @@ package com.arkcronist.enchants.model;
 
 /** A rarity group (SIMPLE, UNIQUE...): colour, name and the book success/destroy ranges it rolls. */
 public record Group(String id, String color, String name, int weight,
-                    int successMin, int successMax, int destroyMin, int destroyMax, int enchanterCost,
+                    double successMin, double successMax, double destroyMin, double destroyMax, int enchanterCost,
                     boolean inEnchanter) {
 
     public boolean curse() {
